Move to compiled babel rather than runtime
