+++ package.json
@@ -3,23 +3,32 @@
   "version": "1.0.0",
   "private": true,
   "scripts": {
-    "test": "./test/runtests.sh"
+    "compile": "babel-compile -c tc-rules/babel src:lib test:.test",
+    "test": "mocha .test/*_test.js",
+    "pretest": "npm run compile",
+    "install": "npm run compile"
   },
   "repository": {
     "type": "git",
     "url": "https://github.com/taskcluster/taskcluster-purge-cache.git"
   },
   "dependencies": {
-    "babel": "^5.0.9",
     "debug": "^2.1.3",
     "lodash": "^3.6.0",
     "promise": "6.1.0",
     "taskcluster-base": "^0.12.2",
-    "taskcluster-client": "^0.23.16"
-  },
-  "devDependencies": {
+    "taskcluster-client": "^0.23.16",
     "mocha": "^2.2.1",
-    "assume": "^1.2.4"
+    "assume": "^1.2.4",
+    "babel-runtime": "^6.6.1",
+    "babel-compile": "^1.0.2",
+    "tc-rules": "^5.0.0",
+    "babel-preset-es2015": "^6.6.0",
+    "babel-plugin-syntax-async-functions": "^6.5.0",
+    "babel-plugin-transform-async-to-generator": "^6.5.0",
+    "babel-plugin-transform-runtime": "^6.6.0",
+    "babel-plugin-transform-strict-mode": "^6.5.2",
+    "babel-plugin-syntax-trailing-function-commas": "^6.5.0"
   },
   "engines": {
     "node": "0.12.x",
