+++ test/mocha.opts
@@ -1,4 +1,3 @@
 --ui tdd
 --timeout 240s
 --reporter spec
---compilers js:babel/register
